refactor(solana): one type gate, drop unused FHE tags
Move host type legality into type_gate.rs, delete Address/Bytes256
from the builder, and pull HostAdmin persist tests out of host_mollusk.
