+++ solana/runtime-tests/tests/host_mollusk.rs
@@ -2434,7 +2434,7 @@ fn mollusk_transaction_later_failure_rolls_back_created_public_output() {
 
 /// Exact HCU cost of `FheExecutionFixture::success_batch`: `Ge` on euint64 operands (152_000) +
 /// `Sub` at euint64 (162_000) + `IfThenElse` at euint64 (55_000). See
-/// `zama-host/src/instructions/fhe_execute/hcu.rs` (EVM `HCULimit` tables).
+/// `zama-host/src/instructions/fhe_execute/hcu/mod.rs` (EVM `HCULimit` tables).
 const FIXTURE_BATCH_HCU: u64 = 152_000 + 162_000 + 55_000; // 369_000
 
 /// Exact HCU cost of the fixture's transient-only execution: a single `Ge` on euint64 operands.
@@ -2562,42 +2562,6 @@ fn set_max_hcu_depth_per_tx_ix(
     )
 }
 
-fn set_host_pause_ix(
-    program_id: Pubkey,
-    admin: Pubkey,
-    host_config: Pubkey,
-    paused: bool,
-) -> Instruction {
-    anchor_ix(
-        program_id,
-        host::accounts::HostAdmin {
-            admin,
-            host_config,
-            event_authority: event_authority(host::id()),
-            program: host::id(),
-        },
-        host::instruction::SetHostPause { paused },
-    )
-}
-
-fn set_grant_deny_list_enabled_ix(
-    program_id: Pubkey,
-    admin: Pubkey,
-    host_config: Pubkey,
-    enabled: bool,
-) -> Instruction {
-    anchor_ix(
-        program_id,
-        host::accounts::HostAdmin {
-            admin,
-            host_config,
-            event_authority: event_authority(host::id()),
-            program: host::id(),
-        },
-        host::instruction::SetGrantDenyListEnabled { enabled },
-    )
-}
-
 fn set_deny_subject_ix(
     program_id: Pubkey,
     payer: Pubkey,
@@ -2619,49 +2583,6 @@ fn set_deny_subject_ix(
             program: host::id(),
         },
         host::instruction::SetDenySubject { subject, denied },
-    )
-}
-
-fn set_admin_ix(
-    program_id: Pubkey,
-    admin: Pubkey,
-    host_config: Pubkey,
-    new_admin: Pubkey,
-) -> Instruction {
-    anchor_ix(
-        program_id,
-        host::accounts::SetAdmin {
-            admin,
-            host_config,
-            new_admin,
-            event_authority: event_authority(host::id()),
-            program: host::id(),
-        },
-        host::instruction::SetAdmin { new_admin },
-    )
-}
-
-fn set_eip712_domain_ix(
-    program_id: Pubkey,
-    admin: Pubkey,
-    host_config: Pubkey,
-    gateway_chain_id: u64,
-    input_verification_contract: [u8; 20],
-    decryption_contract: [u8; 20],
-) -> Instruction {
-    anchor_ix(
-        program_id,
-        host::accounts::HostAdmin {
-            admin,
-            host_config,
-            event_authority: event_authority(host::id()),
-            program: host::id(),
-        },
-        host::instruction::SetEip712Domain {
-            gateway_chain_id,
-            input_verification_contract,
-            decryption_contract,
-        },
     )
 }
 
@@ -2942,42 +2863,6 @@ fn mollusk_set_max_hcu_depth_per_tx_persists_nonzero() {
 }
 
 #[test]
-fn mollusk_set_host_pause_persists() {
-    let program_id = host::id();
-    let admin = Pubkey::new_unique();
-    let (host_config, account) = host_config_account(admin);
-    let context = mollusk_execute_context(admin, vec![(host_config, account)]);
-
-    context.process_and_validate_instruction(
-        &set_host_pause_ix(program_id, admin, host_config, true),
-        &[Check::success()],
-    );
-    assert!(
-        read_host_config(&context, host_config)
-            .expect("config")
-            .paused
-    );
-}
-
-#[test]
-fn mollusk_set_grant_deny_list_enabled_persists() {
-    let program_id = host::id();
-    let admin = Pubkey::new_unique();
-    let (host_config, account) = host_config_account(admin);
-    let context = mollusk_execute_context(admin, vec![(host_config, account)]);
-
-    context.process_and_validate_instruction(
-        &set_grant_deny_list_enabled_ix(program_id, admin, host_config, true),
-        &[Check::success()],
-    );
-    assert!(
-        read_host_config(&context, host_config)
-            .expect("config")
-            .grant_deny_list_enabled
-    );
-}
-
-#[test]
 fn mollusk_set_deny_subject_creates_record() {
     let program_id = host::id();
     let admin = Pubkey::new_unique();
@@ -2996,41 +2881,6 @@ fn mollusk_set_deny_subject_creates_record() {
     let store = context.account_store.borrow();
     let record = store.get(&deny_record).expect("deny record");
     assert_eq!(record.owner, host::id());
-}
-
-#[test]
-fn mollusk_set_admin_transfers_to_pda_without_cosign() {
-    let program_id = host::id();
-    let admin = Pubkey::new_unique();
-    let (new_admin, _) = Pubkey::find_program_address(&[b"new-admin"], &host::id());
-    let (host_config, account) = host_config_account(admin);
-    let context = mollusk_execute_context(admin, vec![(host_config, account)]);
-
-    context.process_and_validate_instruction(
-        &set_admin_ix(program_id, admin, host_config, new_admin),
-        &[Check::success()],
-    );
-    assert_eq!(
-        read_host_config(&context, host_config).expect("config").admin,
-        new_admin
-    );
-}
-
-#[test]
-fn mollusk_set_eip712_domain_persists_zeros() {
-    let program_id = host::id();
-    let admin = Pubkey::new_unique();
-    let (host_config, account) = host_config_account(admin);
-    let context = mollusk_execute_context(admin, vec![(host_config, account)]);
-
-    context.process_and_validate_instruction(
-        &set_eip712_domain_ix(program_id, admin, host_config, 0, [0u8; 20], [0u8; 20]),
-        &[Check::success()],
-    );
-    let config = read_host_config(&context, host_config).expect("config");
-    assert_eq!(config.gateway_chain_id, 0);
-    assert_eq!(config.input_verification_contract, [0u8; 20]);
-    assert_eq!(config.decryption_contract, [0u8; 20]);
 }
 
 #[test]
